This commit is a cleanup of the way we login our users. We no longer load the profiles controller twice upon login, a bug that may have had effects throughout the codebase, since many notifiers were firing two times. I began this commit because I was cleaning up a lot of unused code. Then I noticed that there was a lot of strange ways we were handling the callback we pass to firebase for when a user logs in. Their api is slightly broken, in the sense that  their onAuthStateChanged gives many false positives. Really the function should be called onMaybeAuthStateChanged.
There is still a weird quirk, we need to put a delay of 100ms before
we call FirebaseService.startUi(); If we don't do this then
for some reason the login page does not render.

The other ongoing problem I've seen is that when we login we often don't
have the list of participants display right away. I am not sure if this
is a connection issue.